fix(react-native): implement ArrayBuffer support with custom createVecFromSlice function
- Add create_vec_from_slice function in ffi.rs to safely convert ArrayBuffer to rust::Vec<uint8_t>
- Update bridging-generated.hpp to use createVecFromSlice instead of unsafe memcpy
- Fix ArrayBuffer conversion in React Native example (App.tsx)
- Update backlog: mark ArrayBuffer support as completed with custom implementation details
- Apply changes to both iOS and Android platforms

+++ packages/hwpjs/cpp/CxxHwpjsModule.cpp
@@ -60,7 +60,7 @@ jsi::Value CxxHwpjsModule::fileHeader(jsi::Runtime &rt,
       throw jsi::JSError(rt, "Expected 1 argument");
     }
 
-    auto arg0 = react::bridging::fromJs<rust::Vec<double>>(rt, args[0], callInvoker);
+    auto arg0 = react::bridging::fromJs<rust::Vec<uint8_t>>(rt, args[0], callInvoker);
     auto ret = craby::hwpjs::bridging::fileHeader(*it_, arg0);
 
     return react::bridging::toJs(rt, ret);
@@ -84,7 +84,7 @@ jsi::Value CxxHwpjsModule::toJson(jsi::Runtime &rt,
       throw jsi::JSError(rt, "Expected 1 argument");
     }
 
-    auto arg0 = react::bridging::fromJs<rust::Vec<double>>(rt, args[0], callInvoker);
+    auto arg0 = react::bridging::fromJs<rust::Vec<uint8_t>>(rt, args[0], callInvoker);
     auto ret = craby::hwpjs::bridging::toJson(*it_, arg0);
 
     return react::bridging::toJs(rt, ret);
@@ -108,7 +108,7 @@ jsi::Value CxxHwpjsModule::toMarkdown(jsi::Runtime &rt,
       throw jsi::JSError(rt, "Expected 2 arguments");
     }
 
-    auto arg0 = react::bridging::fromJs<rust::Vec<double>>(rt, args[0], callInvoker);
+    auto arg0 = react::bridging::fromJs<rust::Vec<uint8_t>>(rt, args[0], callInvoker);
     auto arg1 = react::bridging::fromJs<craby::hwpjs::bridging::ToMarkdownOptions>(rt, args[1], callInvoker);
     auto ret = craby::hwpjs::bridging::toMarkdown(*it_, arg0, arg1);
